Move server.rs to connection.rs and change imports.
